Refactor package file logic
