+++ lib/workers/repository.js
@@ -4,6 +4,7 @@ const npmApi = require('../api/npm');
 const defaultsParser = require('../config/defaults');
 const ini = require('ini');
 const packagesWorker = require('./packages');
+const packageFileWorker = require('./package-file');
 const stringify = require('json-stringify-pretty-compact');
 
 module.exports = {
@@ -166,7 +167,7 @@ async function getAllRepoUpgrades(config) {
     delete cascadedConfig.repository;
     delete cascadedConfig.fileName;
     upgrades = upgrades.concat(
-      await packagesWorker.processPackageFile(
+      await packageFileWorker.processPackageFile(
         config.repository,
         packageFile.fileName,
         cascadedConfig
